+++ include/dgl/runtime/ndarray.h
@@ -12,6 +12,7 @@
 #include <utility>
 #include "c_runtime_api.h"
 #include "serializer.h"
+#include "shared_mem.h"
 
 namespace dgl {
 namespace runtime {
@@ -204,46 +205,6 @@ class NDArray {
  * \param tensor The tensor to be saved.
  */
 inline bool SaveDLTensor(dmlc::Stream* strm, const DLTensor* tensor);
-
-#ifndef _WIN32
-/*
- * \brief This class owns shared memory.
- *
- * When the object is gone, the shared memory will also be destroyed.
- * When the shared memory is destroyed, the file corresponding to
- * the shared memory is removed.
- */
-class SharedMemory {
-  bool is_new;
-  std::string name;
-  int fd;
-  void *ptr;
-  size_t size;
-
- public:
-  /*
-   * \brief constructor of the shared memory.
-   * \param name The file corresponding to the shared memory.
-   */
-  explicit SharedMemory(const std::string &name);
-  /*
-   * \brief destructor of the shared memory.
-   * It deallocates the shared memory and removes the corresponding file.
-   */
-  ~SharedMemory();
-  /*
-   * \brief create shared memory.
-   * It creates the file and shared memory.
-   * \param size the size of the shared memory.
-   */
-  void *create_new(size_t size);
-  /*
-   * \brief allocate shared memory that has been created.
-   * \param size the size of the shared memory.
-   */
-  void *open(size_t size);
-};
-#endif  // _WIN32
 
 /*!
  * \brief Reference counted Container object used to back NDArray.
